Doesn't try and destroy what doesn't exist

+++ Multipass/scripts/destroyAll.sh
@@ -1,5 +1,11 @@
 #! /bin/zsh
+FILE_EXISTS=$(multipass list)
+if [ "$FILE_EXISTS" = "No instances found." ]; then
+    echo "None to destroy"
+    exit 0
+fi
 multipass stop --all
 multipass delete --all
 multipass purge
-multipass list
+echo "All gone"
+exit 0
